requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
kombu==5.2.4
mypy-extensions==1.0.0
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
